-moving activity code to lib
