+++ app/src/main/AndroidManifest.xml
@@ -3,20 +3,18 @@
           package="com.tubitv.demo">
     <uses-permission android:name="android.permission.INTERNET"/>
     <application
-        android:name=".DemoApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:largeHeap="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-
         android:theme="@style/AppTheme">
         <activity
             android:name="com.tubitv.demo.MainActivity"
             android:configChanges="keyboard|keyboardHidden|orientation|screenSize|screenLayout|smallestScreenSize|uiMode"
             android:launchMode="singleTop"
-            android:theme="@style/PlayerTheme">
+            android:theme="@style/TubiPlayerTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
