--- app/src/main/java/com/tubitv/demo/DemoApplication.java
@@ -1,33 +1,0 @@
-package com.tubitv.demo;
-
-import android.app.Application;
-
-import com.google.android.exoplayer2.upstream.DataSource;
-import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
-import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
-import com.google.android.exoplayer2.upstream.DefaultHttpDataSourceFactory;
-import com.google.android.exoplayer2.upstream.HttpDataSource;
-import com.google.android.exoplayer2.util.Util;
-
-/**
- * Created by stoyan on 3/21/17.
- */
-public class DemoApplication extends Application {
-
-    protected String userAgent;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        userAgent = Util.getUserAgent(this, "ExoPlayerDemo");
-    }
-
-    public DataSource.Factory buildDataSourceFactory(DefaultBandwidthMeter bandwidthMeter) {
-        return new DefaultDataSourceFactory(this, bandwidthMeter,
-                buildHttpDataSourceFactory(bandwidthMeter));
-    }
-
-    public HttpDataSource.Factory buildHttpDataSourceFactory(DefaultBandwidthMeter bandwidthMeter) {
-        return new DefaultHttpDataSourceFactory(userAgent, bandwidthMeter);
-    }
-}
