+++ lib/build.gradle
@@ -24,6 +24,9 @@ android {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
 
+    compile 'com.android.support:appcompat-v7:25.3.0'
+    compile 'com.android.support.constraint:constraint-layout:1.0.0-alpha9'
+
     compile 'com.google.android.exoplayer:exoplayer:r2.3.0'
 
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
