+++ lib/src/main/AndroidManifest.xml
@@ -5,7 +5,12 @@
         android:allowBackup="true"
         android:label="@string/app_name"
         android:supportsRtl="true">
-
+        <activity
+            android:name="com.tubitv.media.TubiPlayerActivity"
+            android:configChanges="keyboard|keyboardHidden|orientation|screenSize|screenLayout|smallestScreenSize|uiMode"
+            android:launchMode="singleTop"
+            android:theme="@style/TubiPlayerTheme">
+        </activity>
     </application>
 
 </manifest>
