+++ lib/src/main/java/com/tubitv/media/EventLogger.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.tubitv.demo;
+package com.tubitv.media;
 
 import android.os.SystemClock;
 import android.util.Log;
