+++ lib/src/main/java/com/tubitv/media/TubiPlayerActivity.java
@@ -1,0 +1,185 @@
+package com.tubitv.media;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.net.Uri;
+import android.os.Bundle;
+import android.os.Handler;
+import android.text.TextUtils;
+
+import com.google.android.exoplayer2.C;
+import com.google.android.exoplayer2.DefaultLoadControl;
+import com.google.android.exoplayer2.ExoPlayerFactory;
+import com.google.android.exoplayer2.LoadControl;
+import com.google.android.exoplayer2.SimpleExoPlayer;
+import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory;
+import com.google.android.exoplayer2.source.ConcatenatingMediaSource;
+import com.google.android.exoplayer2.source.ExtractorMediaSource;
+import com.google.android.exoplayer2.source.MediaSource;
+import com.google.android.exoplayer2.source.dash.DashMediaSource;
+import com.google.android.exoplayer2.source.dash.DefaultDashChunkSource;
+import com.google.android.exoplayer2.source.hls.HlsMediaSource;
+import com.google.android.exoplayer2.source.smoothstreaming.DefaultSsChunkSource;
+import com.google.android.exoplayer2.source.smoothstreaming.SsMediaSource;
+import com.google.android.exoplayer2.trackselection.AdaptiveTrackSelection;
+import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
+import com.google.android.exoplayer2.trackselection.MappingTrackSelector;
+import com.google.android.exoplayer2.trackselection.TrackSelection;
+import com.google.android.exoplayer2.trackselection.TrackSelector;
+import com.google.android.exoplayer2.ui.SimpleExoPlayerView;
+import com.google.android.exoplayer2.upstream.DataSource;
+import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter;
+import com.google.android.exoplayer2.util.Util;
+
+/**
+ * Created by stoyan on 3/22/17.
+ */
+public class TubiPlayerActivity extends Activity {
+    private SimpleExoPlayer mExoPlayer;
+    private Handler mMainHandler;
+    private SimpleExoPlayerView mSimpleExoPlayerView;
+    private DataSource.Factory mMediaDataSourceFactory;
+    private TrackSelector mTrackSelector;
+    private EventLogger mEventLogger;
+
+    private static final DefaultBandwidthMeter BANDWIDTH_METER = new DefaultBandwidthMeter();
+
+    private boolean shouldAutoPlay;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        shouldAutoPlay = true;
+        mMediaDataSourceFactory = buildDataSourceFactory(true);
+        initLayout();
+
+//        setupExo();
+    }
+
+    @Override
+    public void onNewIntent(Intent intent) {
+        releasePlayer();
+        shouldAutoPlay = true;
+//        clearResumePosition();
+        setIntent(intent);
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        if (Util.SDK_INT > 23) {
+            setupExo();
+        }
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        if ((Util.SDK_INT <= 23 || mExoPlayer == null)) {
+            setupExo();
+        }
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        if (Util.SDK_INT <= 23) {
+            releasePlayer();
+        }
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        if (Util.SDK_INT > 23) {
+            releasePlayer();
+        }
+    }
+
+    private void initLayout() {
+        setContentView(R.layout.activity_tubi_player);
+        mSimpleExoPlayerView = (SimpleExoPlayerView) findViewById(R.id.simple_exoplayer);
+//        mSimpleExoPlayerView.setControllerVisibilityListener(this);
+        mSimpleExoPlayerView.requestFocus();
+    }
+
+    private void setupExo() {
+        // 1. Create a default TrackSelector
+        mMainHandler = new Handler();
+        TrackSelection.Factory videoTrackSelectionFactory =
+                new AdaptiveTrackSelection.Factory(BANDWIDTH_METER);
+        mTrackSelector =
+                new DefaultTrackSelector(videoTrackSelectionFactory);
+
+        // 2. Create a default LoadControl
+        LoadControl loadControl = new DefaultLoadControl();
+
+        // 3. Create the mExoPlayer
+        mExoPlayer =
+                ExoPlayerFactory.newSimpleInstance(this, mTrackSelector, loadControl);
+
+        mEventLogger = new EventLogger((MappingTrackSelector) mTrackSelector);
+        mExoPlayer.addListener(mEventLogger);
+        mExoPlayer.setAudioDebugListener(mEventLogger);
+        mExoPlayer.setVideoDebugListener(mEventLogger);
+        mExoPlayer.setMetadataOutput(mEventLogger);
+
+        mSimpleExoPlayerView.setPlayer(mExoPlayer);
+        mExoPlayer.setPlayWhenReady(shouldAutoPlay);
+
+        //fake media
+        Uri[] uris = new Uri[1];
+        String[] extensions = new String[1];
+//        uris[0] = Uri.parse("http://www.youtube.com/api/manifest/dash/id/bf5bb2419360daf1/source/youtube?as=fmp4_audio_clear,fmp4_sd_hd_clear&sparams=ip,ipbits,expire,source,id,as&ip=0.0.0.0&ipbits=0&expire=19000000000&signature=51AF5F39AB0CEC3E5497CD9C900EBFEAECCCB5C7.8506521BFC350652163895D4C26DEE124209AA9E&key=ik0");
+        uris[0] = Uri.parse("http://c11.adrise.tv/v2/sources/content-owners/marvista/342067/v201701141225-1280x714-,439,1099,1714,2062,2943,k.mp4.m3u8?csQ-b00i5q5qYJ0FgNlUslZ35UTdh4_GBHHAGq_-CMH04cDQXJQ76NFQK2pYpj87");
+//        uris[0] = Uri.parse("http://c11.adrise.tv/v2/sources/content-owners/marvista/342067/v201701141225-1280x714-2943k.mp4.m3u8?csQ-b00i5q5qYJ0FgNlUslZ35UTdh4_GBHHAGq_-CMH04cDQXJQ76NFQK2pYpj87");
+        extensions[0] = "m3u8";
+//        extensions[0] = C.TYPE_HLS;
+        MediaSource[] mediaSources = new MediaSource[uris.length];
+        mediaSources[0] = buildMediaSource(uris[0], extensions[0]);
+        MediaSource mediaSource = mediaSources.length == 1 ? mediaSources[0]
+                : new ConcatenatingMediaSource(mediaSources);
+        mExoPlayer.prepare(mediaSource, true, false);
+    }
+
+    private void releasePlayer() {
+        if (mExoPlayer != null) {
+            shouldAutoPlay = mExoPlayer.getPlayWhenReady();
+            mExoPlayer.release();
+            mExoPlayer = null;
+            mTrackSelector = null;
+        }
+    }
+
+    private MediaSource buildMediaSource(Uri uri, String overrideExtension) {
+        int type = TextUtils.isEmpty(overrideExtension) ? Util.inferContentType(uri)
+                : Util.inferContentType("." + overrideExtension);
+        switch (type) {
+            case C.TYPE_SS:
+                return new SsMediaSource(uri, buildDataSourceFactory(false),
+                        new DefaultSsChunkSource.Factory(mMediaDataSourceFactory), mMainHandler, mEventLogger);
+            case C.TYPE_DASH:
+                return new DashMediaSource(uri, buildDataSourceFactory(false),
+                        new DefaultDashChunkSource.Factory(mMediaDataSourceFactory), mMainHandler, mEventLogger);
+            case C.TYPE_HLS:
+                return new HlsMediaSource(uri, mMediaDataSourceFactory, mMainHandler, mEventLogger);
+            case C.TYPE_OTHER:
+                return new ExtractorMediaSource(uri, mMediaDataSourceFactory, new DefaultExtractorsFactory(),
+                        mMainHandler, mEventLogger);
+            default: {
+                throw new IllegalStateException("Unsupported type: " + type);
+            }
+        }
+    }
+
+    /**
+     * Returns a new DataSource factory.
+     *
+     * @param useBandwidthMeter Whether to set {@link #BANDWIDTH_METER} as a listener to the new
+     *                          DataSource factory.
+     * @return A new DataSource factory.
+     */
+    private DataSource.Factory buildDataSourceFactory(boolean useBandwidthMeter) {
+        return MediaHelper.buildDataSourceFactory(this, useBandwidthMeter ? BANDWIDTH_METER : null);
+    }
+}
