+++ lib/src/main/res/layout/activity_tubi_player.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.tubitv.demo.MainActivity">
+    tools:context="com.tubitv.media.TubiPlayerActivity">
 
     <com.google.android.exoplayer2.ui.SimpleExoPlayerView android:id="@+id/simple_exoplayer"
         android:layout_width="wrap_content"
